requirements.txt
torch==1.17.1
Pillow==9.4.0
torchvision==0.14.1